More efficient and maintainable partitioning of OsmoGridModels Division of functionality into several objects Adapting tests
