+++ src/main/scala/edu/ie3/osmogrid/lv/coordinator/LvCoordinator.scala
@@ -13,7 +13,7 @@ import edu.ie3.osmogrid.cfg.OsmoGridConfig
 import edu.ie3.osmogrid.io.input.BoundaryAdminLevel
 import edu.ie3.osmogrid.io.input.InputDataProvider
 import edu.ie3.osmogrid.io.input.InputDataProvider.{ReqAssetTypes, ReqOsm}
-import edu.ie3.osmogrid.lv.LvRegionCoordinator
+import edu.ie3.osmogrid.lv.region_coordinator.LvRegionCoordinator
 import edu.ie3.osmogrid.model.SourceFilter.LvFilter
 
 import java.util.UUID
