+++ src/test/resources/edu/ie3/osmogrid/lv/region_coordinator/DoBoCas.pbf
