+++ src/test/resources/edu/ie3/osmogrid/lv/region_coordinator/DoBoCas_boundaries_level8.csv
